authentication

+++ src/app/controllers/ProjectController.js
@@ -3,61 +3,63 @@ const pool = require('../../config/db');
 class ProjectController {
     async create(req, res) {
         try {
-            const { name, description, location, user_id, start_date, end_date, quantity, uni_ids } = req.body;
+            const { name, description, location, start_date, end_date, quantity, uni_id } = req.body;
+
+            const user_id = req.userID;
+
             if (!name) {
                 return res.status(400).json('Invalid Project');
             }
 
-            const query = `INSERT INTO project (name, description, location, user_id, start_date, end_date, quantity) 
-                    VALUES ($1, $2, $3, $4, $5, $6, $7)
+            const query = `INSERT INTO project (name, description, location,start_date, end_date, quantity, uni_id,  user_id) 
+                    VALUES ($1, $2, $3, $4, $5, $6, $7, $8)
                     RETURNING id
             `;
 
-            const projectUniQuery = `INSERT INTO project_uni (project_id, uni_id, is_checked) VALUES ($1, $2, $3)`;
+            // const projectUniQuery = `INSERT INTO project_uni (project_id, uni_id, is_checked) VALUES ($1, $2, $3)`;
 
-            for (const uni_id of uni_ids) {
-                try {
-                    const resUni = await pool.query('SELECT * from university where id = $1', [uni_id]);
-                    if (resUni.rows.length === 0) {
-                        return res.status(404).json({ message: `Univeristy with id = ${uni_id} is not existed.` });
-                    }
-                } catch (error) {
-                    return res.status(404).json({ message: `Error with uni_id = ${uni_id}` });
+            try {
+                const resUni = await pool.query('SELECT * from university where id = $1', [uni_id]);
+                if (resUni.rows.length === 0) {
+                    return res.status(404).json({ message: `Univeristy with id = ${uni_id} is not existed.` });
                 }
+            } catch (error) {
+                return res.status(404).json({ message: `Error with uni_id = ${uni_id}` });
             }
 
             const response = await pool.query(query, [
                 name,
                 description,
                 location,
-                user_id,
                 start_date,
                 end_date,
                 quantity,
+                uni_id,
+                user_id,
             ]);
 
-            for (const uni_id of uni_ids) {
-                try {
-                    const createProjectUni = await pool.query(projectUniQuery, [response.rows[0].id, uni_id, null]);
-                } catch (error) {
-                    return res.status(404).json({ message: `Error with uni_id = ${uni_id}` });
-                }
-            }
+            // try {
+            //     const createProjectUni = await pool.query(projectUniQuery, [response.rows[0].id, uni_id, null]);
+            // } catch (error) {
+            //     return res.status(404).json({ message: `Error with uni_id = ${uni_id}` });
+            // }
 
             return res.status(201).json({
                 message: 'Project created successfully',
                 data: {
+                    id: response.rows[0].id,
                     name,
                     description,
                     location,
-                    user_id,
                     start_date,
                     end_date,
                     quantity,
-                    id: response.rows[0].id,
+                    uni_id,
+                    user_id,
                 },
             });
         } catch (error) {
+            console.log(error);
             return res.status(500).json('Internal Server Error');
         }
     }
@@ -100,12 +102,18 @@ class ProjectController {
     async update(req, res) {
         try {
             let { id, name, description, location, user_id, start_date, end_date, quantity } = req.body;
+
             const prevProject = await pool.query('select * from project where id = $1', [id]);
+
             if (prevProject.rows.length === 0) return res.status(404).send({ message: 'Project not found' });
+
+            if (prevProject.rows[0].user_id !== req.userID)
+                return res.status(404).send({ message: "That project wasn't created by you." });
+
             const query = `
                 UPDATE project
-                SET name = $1, description = $2, location = $3, user_id = $4, start_date = $5, end_date = $6, quantity = $7
-                WHERE id = $8
+                SET name = $1, description = $2, location = $3, start_date = $4, end_date = $5, quantity = $6
+                WHERE id = $7
             `;
 
             if (!name) name = prevProject.rows[0].name;
@@ -117,7 +125,7 @@ class ProjectController {
             if (!quantity) quantity = prevProject.rows[0].quantity;
             if (!id) id = prevProject.rows[0].id;
 
-            const values = [name, description, location, user_id, start_date, end_date, quantity, id];
+            const values = [name, description, location, start_date, end_date, quantity, id];
             const response = await pool.query(query, values);
 
             return res.status(200).json({
@@ -134,8 +142,8 @@ class ProjectController {
         try {
             const id = parseInt(req.params.slug);
             const isChecked = req.body.isChecked;
-
             const resProject = await pool.query('SELECT * from project WHERE id = $1', [id]);
+
             if (resProject.rows.length === 0) {
                 return res.status(404).json({ message: 'Project not found!' });
             }
@@ -144,6 +152,11 @@ class ProjectController {
                 return res.status(400).json({ message: 'The project has been handle' });
             }
 
+            const userUniId = (await pool.query('SELECT * from users WHERE id = $1', [req.userID])).rows[0].uni_id;
+
+            if (userUniId !== resProject.rows[0].uni_id) {
+                return res.status(400).json({ message: "That project wasn't created for you." });
+            }
 
             const query = `
                 UPDATE project SET is_checked = $1 WHERE id = $2
@@ -152,12 +165,12 @@ class ProjectController {
             const values = [isChecked, id];
             const response = await pool.query(query, values);
 
-
             return res.status(200).json({
                 message: 'Update Project successfully!',
-                data: {...resProject.rows[0], is_checked: req.body.isChecked },
+                data: { ...resProject.rows[0], is_checked: req.body.isChecked },
             });
         } catch (err) {
+            console.log(err)
             return res.status(500).json('Internal Server Error');
         }
     }
@@ -168,10 +181,17 @@ class ProjectController {
 
             const resProject = await pool.query('SELECT * from project where id = $1', [id]);
 
+            if (resProject.rows.length === 0) return res.status(404).send({ message: 'Project not found' });
+
             if (resProject.rows.length === 0) {
                 return res.status(400).json({ message: 'Project not found' });
             }
 
+            if (resProject.rows[0].user_id !== req.userID) {
+                return res.status(404).send({ message: "That project wasn't created by you." });
+            }
+
+            
             const deleteProjectQuery = `
                 DELETE FROM project
                 WHERE id = $1
@@ -205,11 +225,11 @@ class ProjectController {
                         university.id as uni_id
                         FROM project_uni
                 JOIN project ON project_uni.project_id = project.id
-                JOIN university ON project_uni.uni_id = university.id`
+                JOIN university ON project_uni.uni_id = university.id`,
             );
             return res.status(200).json({
-                data: response.rows
-            })
+                data: response.rows,
+            });
         } catch (error) {
             console.log(error);
             return res.status(500).json('Internal Server Error');
